+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.95] - 2024-07-14
+### Added
+* `SetComponentInNormalizedDirection` and `GetComponentInNormalizedDirection` to `Vector3Utils`
+
 ## [0.0.94] - 2024-07-03
 ### Added
 * `Vector3IntExtensions.ToVector3`
+* overloaded `QuaternionExtensions.ToDirectionVector`
+* `Vector3.Utils.SetComponentInDirection`
 
 ## [0.0.93] - 2024-06-28
 ### Added